feat: Modify env parameter (#333)
When `devParameter` is not set, it is difficult to understand the behavior of whether the CDK defaults are used or not, so env was modified.
Also, removed the default region because it's not be needed.

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [MIT No Attribution (MIT-0)].

[MIT No Attribution (MIT-0)]: https://github.com/aws/mit-0