fix: deprecated code, RDS's instanceProps and CloudWatch Synthetics's older runtime (#662)
#604 , #652 show the issues of deprecated properties.
This PR is to fix them.

In #604, `instanceProps` was deprecated.
So, it uses `writer` and `readers` props instead of it.

In #652, `synthetics.Runtime.SYNTHETICS_NODEJS_PUPPETEER_3_5` was deprecated.
So, it updates `puppeteer` version from `3.5` to `6.0`.
Because BLEA`s comment recommends to use the latest version.


---

By submitting this pull request, I confirm that my contribution is made under the terms of the [MIT No Attribution (MIT-0)].

[MIT No Attribution (MIT-0)]: https://github.com/aws/mit-0